feat(src): move `parseManifestFromStdin` function from if-diff helpers to common helpers file
