+++ src/common/util/helpers.ts
@@ -1,3 +1,4 @@
+import {createInterface} from 'node:readline/promises';
 import {exec} from 'child_process';
 import * as path from 'path';
 import {promisify} from 'util';
@@ -18,4 +19,47 @@ export const prependFullFilePath = (filePath: string) => {
   }
 
   return path.normalize(`${processRunningPath}/${filePath}`);
+};
+
+/**
+ * Checks if there is data piped, then collects it.
+ * Otherwise returns empty string.
+ */
+const collectPipedData = async () => {
+  if (process.stdin.isTTY) {
+    return '';
+  }
+
+  const readline = createInterface({
+    input: process.stdin,
+  });
+
+  const data: string[] = [];
+
+  for await (const line of readline) {
+    data.push(line);
+  }
+
+  return data.join('\n');
+};
+
+/**
+ * Checks if there is piped data, tries to parse yaml from it.
+ * Returns empty string if haven't found anything.
+ */
+export const parseManifestFromStdin = async () => {
+  const pipedSourceManifest = await collectPipedData();
+
+  if (!pipedSourceManifest) {
+    return '';
+  }
+
+  const regex = /# start((?:.*\n)+?)# end/;
+  const match = regex.exec(pipedSourceManifest);
+
+  if (!match) {
+    return '';
+  }
+
+  return match![1];
 };
