+++ src/if-diff/util/helpers.ts
@@ -1,4 +1,3 @@
-import {createInterface} from 'node:readline/promises';
 import {Difference} from '../types/compare';
 
 /**
@@ -82,47 +81,4 @@ export const formatNotMatchingLog = (message: Difference) => {
       console.log(`${key}: ${message[key]}`);
     }
   });
-};
-
-/**
- * Checks if there is data piped, then collects it.
- * Otherwise returns empty string.
- */
-const collectPipedData = async () => {
-  if (process.stdin.isTTY) {
-    return '';
-  }
-
-  const readline = createInterface({
-    input: process.stdin,
-  });
-
-  const data: string[] = [];
-
-  for await (const line of readline) {
-    data.push(line);
-  }
-
-  return data.join('\n');
-};
-
-/**
- * Checks if there is piped data, tries to parse yaml from it.
- * Returns empty string if haven't found anything.
- */
-export const parseManifestFromStdin = async () => {
-  const pipedSourceManifest = await collectPipedData();
-
-  if (!pipedSourceManifest) {
-    return '';
-  }
-
-  const regex = /# start((?:.*\n)+?)# end/;
-  const match = regex.exec(pipedSourceManifest);
-
-  if (!match) {
-    return '';
-  }
-
-  return match![1];
 };
